R6 master class for storing information that is available across all modules.

app_master is the master class used across the shiny app. It Stores all tibbles that are then used by various mods. These tibbles are stored in reactiveValues rvals

Public fields

params

initialization parameters

rvals

reactiveValues instance that stores all datasets needed in app

svals

non-reactive static datasets needed in the application. Note you can only pre-load these

Methods

Public methods

app_master$new()

app_master$preload_master()

app_master$preload_master_with_config()

app_master$mdata()

app_master$dataset_names()

app_master$data_by_name()

app_master$data_by_index()

app_master$clone()

Method new()

Standard R6 Initialize function

Usage

app_master$new(params)

Arguments

params

the config yml driven params for initialization

Returns

a new app_master object Preload app_master with master_data `

Note the mdata value in reactiveValues (rvals) will be overwritten by what is provided

Method preload_master()

Usage

app_master$preload_master(master_data)

Arguments

master_data

mdata in rvals will be raplced with the mmaster_data provided Preload app_master with csv files provided in config `Note this creates new mdata overiding rvals

Method preload_master_with_config()

Usage

app_master$preload_master_with_config()

Returns

self object access mdata easily

Method mdata()

Usage

app_master$mdata()

Returns

the mdata object in reactiveValues access dataset names as list

Method dataset_names()

Usage

app_master$dataset_names()

Returns

the names of datasets search for a tibble based on dataset_name

Method data_by_name()

Usage

app_master$data_by_name(dataset_name)

Arguments

dataset_name

the name of the dataset to lookup

Returns

the mapped dataset in tibble format search for a tibble based on index in mdata

Method data_by_index()

Usage

app_master$data_by_index(index)

Arguments

index

the row index of the dataset

Returns

the mapped dataset in tibble format

Method clone()

The objects of this class are cloneable with this method.

Usage

app_master$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.